08-financial-analysis-agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
